Finished Service User Ctrl

+++ backend/src/main/java/org/MoneyManagement/backend/service/impl/UserServiceImpl.java
@@ -1,6 +1,7 @@
 package org.MoneyManagement.backend.service.impl;
 
 import org.MoneyManagement.backend.entity.User;
+import org.MoneyManagement.backend.model.GeneralUserRequest;
 import org.MoneyManagement.backend.repository.UserDao;
 import org.MoneyManagement.backend.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -12,6 +13,8 @@ import java.util.List;
 @Service
 public class UserServiceImpl implements UserService {
 
+    enum requestType {ChangeName, ChangePassword, ChangeUserName, UndefinedRequest}
+
     @Autowired
     private UserDao userDao;
 
@@ -22,33 +25,68 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public void addUser(GeneralUserRequest request) {
+    public User addUser(GeneralUserRequest request) {
+
         User tempUser = new User();
-        tempUser.setId(request.getId());
-        tempUser.setName(request.getName());
-        tempUser.setPassword(request.getPassword());
-        tempUser.setCreatedAt(new Date());
-        tempUser.setUserName(request.getUserName());
-        if(tempUser.getId() != null)
-        {
-            userDao.save(tempUser);
+        User tempSavedUser = new User();
+
+        if (request != null) {
+            tempUser.setName(request.getName());
+            tempUser.setPassword(request.getPassword());
+            tempUser.setCreatedAt(new Date());
+            tempUser.setUserName(request.getUserName());
+            tempSavedUser = userDao.save(tempUser);
+
+            if (tempUser.equalsNew(tempSavedUser)) {
+                /*Nothing to do, user saved correctyl*/
+            } else {
+                /*ERROR, user not saved correctly*/
+            }
+
+        } else {
+            /*AleTau: Handle null request*/
+        }
+        return tempSavedUser;
+    }
+
+    @Override
+    public void removeUserById(String id) {
+        if (id != null) {
+            userDao.deleteById(id);
+        } else {
+            /*AleTau: Handle id = null exception*/
         }
     }
 
     @Override
-    public void removeUserById(String id)
-    {
-        userDao.deleteById(id);
-    }
+    public User editUser(GeneralUserRequest request) {
+        User tempUser = new User();
+        User tempUpdatedUser = new User;
+        String userId = request.getId();
 
-    @Override
-    public void editUser(GeneralUserRequest request) {
+        if (userId != null) {
+            tempUser = userDao.findById(userId);
+            tempUser.setName(request.getName());
+            tempUser.setPassword(request.getPassword());
+            tempUser.setUserName(request.getUserName());
 
+            tempUpdatedUser = userDao.save(tempUser);
+
+            if(tempUser.equalsNew(tempUpdatedUser) == true )
+            {
+                /*User saved correctly, Nothing to do*/
+            } else {
+                /*Handle user not saved correctly*/
+            }
+        } else {
+            /*AleTau: Handle id = null exception*/
+        }
+    return tempUpdatedUser;
     }
 
     @Override
     public User getUserbyId(String id) {
-        return null;
+        return userDao.findById(id);
     }
 
     @Override
